refactor(schedule): unify calendar and list views into shared components
Extract ScheduleCalendarView, ScheduleListView, and schedule-utils into
shared files used by both the org schedule page and the agent schedule
card. Add Run now and row navigation (onOpenDetails) to the agent
schedule card so team and page views behave identically — the only
difference is the schedule page shows an agent column for multi-agent.
