+++ turbo/apps/platform/src/views/zero-page/zero-schedule-card.tsx
@@ -15,17 +15,8 @@ import {
   setSaveError$,
   togglingIds$,
   setTogglingIds$,
-  calendarPopoverEntryId$,
-  setCalendarPopoverEntryId$,
 } from "../../signals/zero-page/schedule-card.ts";
-import {
-  IconPlus,
-  IconList,
-  IconLayoutGrid,
-  IconPencil,
-  IconTrash,
-} from "@tabler/icons-react";
-import { LoadingSwitch } from "../components/loading-switch.tsx";
+import { IconPlus, IconList, IconLayoutGrid } from "@tabler/icons-react";
 import {
   Card,
   CardContent,
@@ -33,19 +24,14 @@ import {
   Tabs,
   TabsList,
   TabsTrigger,
-  cn,
 } from "@vm0/ui";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@vm0/ui/components/ui/popover";
 import { throwIfAbort, detach, Reason } from "../../signals/utils.ts";
 import {
   ScheduleFormDialog,
   type ScheduleFormValues,
 } from "./schedule-dialog.tsx";
-import emptyScheduleImg from "./assets/empty-schedule.webp";
+import { ScheduleCalendarView } from "./schedule-calendar-view.tsx";
+import { ScheduleListView } from "./schedule-list-view.tsx";
 import {
   Dialog,
   DialogContent,
@@ -55,88 +41,9 @@ import {
   DialogFooter,
 } from "@vm0/ui/components/ui/dialog";
 
-export const WEEKDAY_LABELS = [
-  "Mon",
-  "Tue",
-  "Wed",
-  "Thu",
-  "Fri",
-  "Sat",
-  "Sun",
-] as const;
-const CALENDAR_TIME_SLOTS = [
-  "6:00 AM",
-  "9:00 AM",
-  "12:00 PM",
-  "6:00 PM",
-] as const;
+import type { ScheduleEntry } from "./schedule-utils";
 
-export interface ScheduleEntry {
-  id: string;
-  time: string;
-  prompt: string;
-  description?: string | null;
-  /** Schedule name used for API operations (edit/delete). */
-  name?: string;
-  enabled?: boolean;
-  notifyEmail?: boolean;
-  notifySlack?: boolean;
-  /** IANA timezone from the server (not derivable from `time` alone). */
-  timezone?: string;
-  slackChannelId?: string | null;
-  /** Raw interval in seconds for loop schedules */
-  intervalSeconds?: number | null;
-}
-
-function DeleteButton({
-  name,
-  label,
-  onDelete,
-}: {
-  name: string;
-  label: string;
-  onDelete: (name: string) => Promise<void>;
-}) {
-  const [open, setOpen] = useState(false);
-  return (
-    <>
-      <button
-        type="button"
-        onClick={() => setOpen(true)}
-        className="rounded p-1.5 text-muted-foreground hover:bg-muted hover:text-destructive focus:outline-none focus-visible:ring-2 focus-visible:ring-ring shrink-0"
-        aria-label={`Delete ${label}`}
-      >
-        <IconTrash size={14} stroke={1.5} />
-      </button>
-      <Dialog open={open} onOpenChange={setOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Delete schedule?</DialogTitle>
-            <DialogDescription>
-              This will permanently delete the schedule{" "}
-              <span className="font-medium text-foreground">{name}</span>. This
-              action cannot be undone.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setOpen(false)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={() => {
-                setOpen(false);
-                detach(onDelete(name), Reason.DomCallback);
-              }}
-            >
-              Delete
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-    </>
-  );
-}
+export { WEEKDAY_LABELS, type ScheduleEntry } from "./schedule-utils";
 
 function formatTimeOfDay(hour: number, minute: number): string {
   if (hour === 0 && minute === 0) {
@@ -298,165 +205,6 @@ export function parseScheduleTimeString(timeStr: string): ParsedScheduleTime {
   return defaultParsed({ hour, minute });
 }
 
-function parseScheduleTime(timeStr: string): {
-  dayIndices: number[];
-  timeLabel: string;
-} {
-  if (timeStr.match(/Every \d+ (minutes?|seconds?)/) || timeStr === "Now") {
-    return { dayIndices: [], timeLabel: "" };
-  }
-  const match = timeStr.match(/at (\d{1,2}:\d{2} (?:AM|PM))$/);
-  const timeLabel = match ? match[1] : "9:00 AM";
-  if (timeStr.startsWith("Every day") && !timeStr.startsWith("Every weekday")) {
-    return { dayIndices: [0, 1, 2, 3, 4, 5, 6], timeLabel };
-  }
-  if (timeStr.startsWith("Every weekday")) {
-    return { dayIndices: [0, 1, 2, 3, 4], timeLabel };
-  }
-  const dayMap: Record<string, number> = {
-    Monday: 0,
-    Tuesday: 1,
-    Wednesday: 2,
-    Thursday: 3,
-    Friday: 4,
-    Saturday: 5,
-    Sunday: 6,
-  };
-  const onMatch = timeStr.match(
-    /on ((?:(?:Monday|Tuesday|Wednesday|Thursday|Friday|Saturday|Sunday)(?:,\s*)?)+) at/,
-  );
-  if (onMatch) {
-    const names = onMatch[1].split(/,\s*/);
-    const indices = names.map((n) => dayMap[n] ?? -1).filter((i) => i >= 0);
-    if (indices.length > 0) {
-      return { dayIndices: indices, timeLabel };
-    }
-  }
-  if (timeStr.startsWith("Every week")) {
-    return { dayIndices: [0, 1, 2, 3, 4, 5, 6], timeLabel };
-  }
-  if (timeStr.startsWith("Every month")) {
-    return { dayIndices: [], timeLabel: "" };
-  }
-  return { dayIndices: [], timeLabel };
-}
-
-/**
- * Convert a time label like "9:00 AM" to minutes since midnight.
- */
-function timeLabelToMinutes(label: string): number {
-  const match = label.match(/^(\d{1,2}):(\d{2})\s*(AM|PM)$/);
-  if (!match) {
-    return 0;
-  }
-  let hour = Number(match[1]);
-  const minute = Number(match[2]);
-  const ampm = match[3];
-  if (ampm === "PM" && hour !== 12) {
-    hour += 12;
-  }
-  if (ampm === "AM" && hour === 12) {
-    hour = 0;
-  }
-  return hour * 60 + minute;
-}
-
-/**
- * Build the calendar time slots by merging default slots with entry-specific times,
- * sorted chronologically.
- */
-export function buildCalendarTimeSlots(
-  scheduleList: readonly Readonly<ScheduleEntry>[],
-): string[] {
-  const slotSet = new Set<string>(CALENDAR_TIME_SLOTS);
-  for (const entry of scheduleList) {
-    if (entry.enabled === false) {
-      continue;
-    }
-    const { timeLabel } = parseScheduleTime(entry.time);
-    if (timeLabel) {
-      slotSet.add(timeLabel);
-    }
-  }
-  return [...slotSet].sort(
-    (a, b) => timeLabelToMinutes(a) - timeLabelToMinutes(b),
-  );
-}
-
-export function getEntriesInCell(
-  scheduleList: ScheduleEntry[],
-  dayIndex: number,
-  timeLabel: string,
-): ScheduleEntry[] {
-  return scheduleList.filter((entry) => {
-    if (entry.enabled === false) {
-      return false;
-    }
-    const { dayIndices, timeLabel: t } = parseScheduleTime(entry.time);
-    return t === timeLabel && dayIndices.includes(dayIndex);
-  });
-}
-
-function CalendarEntryPopover({
-  entry,
-  cellKey,
-  onEdit,
-}: {
-  entry: ScheduleEntry;
-  cellKey: string;
-  onEdit: (entry: ScheduleEntry) => void;
-}) {
-  const hoveredId = useGet(calendarPopoverEntryId$);
-  const setHoveredId = useSet(setCalendarPopoverEntryId$);
-  const popoverId = `${entry.id}-${cellKey}`;
-  const open = hoveredId === popoverId;
-  const setOpen = (v: boolean) => setHoveredId(v ? popoverId : null);
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          onMouseEnter={() => setOpen(true)}
-          onMouseLeave={() => setOpen(false)}
-          onDoubleClick={() => onEdit(entry)}
-          className="w-full min-h-0 rounded px-1.5 py-0.5 text-[11px] leading-tight line-clamp-2 break-words border border-blue-700/40 bg-blue-700/15 text-blue-800 hover:bg-blue-700/25 focus:outline-none focus-visible:ring-2 focus-visible:ring-blue-600/50 dark:text-blue-200 dark:border-blue-600/40 dark:bg-blue-900/25 dark:hover:bg-blue-900/35 text-left"
-          aria-label={`${entry.time}: ${entry.description || entry.prompt}`}
-        >
-          {entry.description || entry.prompt}
-        </button>
-      </PopoverTrigger>
-      <PopoverContent
-        align="start"
-        sideOffset={0}
-        className="w-80 p-3 flex flex-col gap-3"
-        onMouseEnter={() => setOpen(true)}
-        onMouseLeave={() => setOpen(false)}
-      >
-        <div className="relative flex flex-col gap-1.5 pr-8">
-          <div className="absolute top-0 right-0">
-            <button
-              type="button"
-              onClick={() => onEdit(entry)}
-              className="rounded p-1 text-muted-foreground hover:bg-muted hover:text-foreground transition-colors"
-              aria-label={`Edit ${entry.time}`}
-            >
-              <IconPencil size={14} stroke={1.5} />
-            </button>
-          </div>
-          <p className="text-xs text-muted-foreground">{entry.time}</p>
-          {entry.description && (
-            <p className="text-sm font-medium text-foreground leading-snug">
-              {entry.description}
-            </p>
-          )}
-          <p className="text-sm text-foreground leading-snug">{entry.prompt}</p>
-        </div>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
 interface ZeroScheduleCardProps {
   title: string;
   subtitle: string;
@@ -485,6 +233,10 @@ interface ZeroScheduleCardProps {
     name: string;
     enabled: boolean;
   }) => Promise<void>;
+  /** When provided, called to trigger an immediate run for a schedule. */
+  onRunNow?: (entry: ScheduleEntry) => Promise<void>;
+  /** When provided, row clicks navigate to the schedule detail. */
+  onOpenDetails?: (entry: ScheduleEntry) => void;
   /** When true, the save button shows a loading state. */
   saving?: boolean;
   /** Default timezone for new schedules. Falls back to browser timezone. */
@@ -498,6 +250,8 @@ export function ZeroScheduleCard({
   onSave,
   onDelete,
   onToggleEnabled,
+  onRunNow,
+  onOpenDetails,
   saving,
 }: ZeroScheduleCardProps) {
   const scheduleViewMode = useGet(scheduleViewMode$);
@@ -524,9 +278,66 @@ export function ZeroScheduleCard({
     setAddScheduleOpen(true);
   };
 
+  const [pendingDelete, setPendingDelete] = useState<ScheduleEntry | null>(
+    null,
+  );
+
   const openEditSchedule = (entry: ScheduleEntry) => {
     setSaveError(null);
     setEditingScheduleId(entry.id);
+  };
+
+  const handleToggle = onToggleEnabled
+    ? (entry: ScheduleEntry, enabled: boolean) => {
+        if (entry.name === undefined) {
+          return;
+        }
+        const id = entry.id;
+        setTogglingIds((prev) => new Set([...prev, id]));
+        detach(
+          onToggleEnabled({ name: entry.name, enabled }).finally(() => {
+            setTogglingIds((prev) => {
+              const next = new Set(prev);
+              next.delete(id);
+              return next;
+            });
+          }),
+          Reason.DomCallback,
+        );
+      }
+    : undefined;
+
+  const handleDelete = onDelete
+    ? (entry: ScheduleEntry) => {
+        setPendingDelete(entry);
+      }
+    : undefined;
+
+  const [runningIds, setRunningIds] = useState<Set<string>>(new Set());
+
+  const handleRunNow = onRunNow
+    ? async (entry: ScheduleEntry) => {
+        const id = entry.id;
+        setRunningIds((prev) => new Set([...prev, id]));
+        try {
+          await onRunNow(entry);
+        } finally {
+          setRunningIds((prev) => {
+            const next = new Set(prev);
+            next.delete(id);
+            return next;
+          });
+        }
+      }
+    : undefined;
+
+  const confirmDelete = () => {
+    const entry = pendingDelete;
+    if (!entry?.name || !onDelete) {
+      return;
+    }
+    setPendingDelete(null);
+    detach(onDelete(entry.name), Reason.DomCallback);
   };
 
   const handleCreateSave = (values: ScheduleFormValues) => {
@@ -690,243 +501,30 @@ export function ZeroScheduleCard({
         </header>
 
         {scheduleViewMode === "list" && (
-          <section className="flex flex-col">
-            {scheduleList.length === 0 ? (
-              <div className="flex flex-col items-center justify-center py-12 gap-3">
-                <img
-                  src={emptyScheduleImg}
-                  alt="No entries"
-                  loading="lazy"
-                  className="h-20 w-20 object-contain opacity-80"
-                />
-                <div className="text-center">
-                  <p className="text-sm font-medium text-foreground">
-                    No runs scheduled
-                  </p>
-                  <p className="text-xs text-muted-foreground mt-1">
-                    Pick a time and tell your agent what to do.
-                  </p>
-                </div>
-              </div>
-            ) : (
-              <ul className="flex flex-col" role="list">
-                {scheduleList.map((entry) => {
-                  const toggling = togglingIds.has(entry.id);
-                  return (
-                    <li
-                      key={entry.id}
-                      className="flex items-center gap-3 py-2.5 border-b border-border/50 last:border-b-0 text-sm text-foreground hover:bg-muted/30 -mx-1 px-1 rounded transition-colors"
-                    >
-                      {onToggleEnabled && entry.name !== undefined && (
-                        <LoadingSwitch
-                          checked={entry.enabled !== false}
-                          loading={toggling}
-                          onCheckedChange={(checked) => {
-                            const id = entry.id;
-                            const name = entry.name;
-                            if (name === undefined) {
-                              return;
-                            }
-                            setTogglingIds((prev) => new Set([...prev, id]));
-                            detach(
-                              onToggleEnabled({
-                                name,
-                                enabled: checked,
-                              }).finally(() => {
-                                setTogglingIds((prev) => {
-                                  const next = new Set(prev);
-                                  next.delete(id);
-                                  return next;
-                                });
-                              }),
-                              Reason.DomCallback,
-                            );
-                          }}
-                          ariaLabel={`${entry.enabled !== false ? "Disable" : "Enable"} ${entry.time}`}
-                        />
-                      )}
-                      <span
-                        className={cn(
-                          "min-w-0 shrink-0",
-                          entry.enabled === false && "text-muted-foreground",
-                        )}
-                      >
-                        {entry.time}
-                      </span>
-                      <span className="min-w-0 flex-1 text-muted-foreground text-xs truncate">
-                        {entry.description || entry.prompt}
-                      </span>
-                      <button
-                        type="button"
-                        onClick={() => openEditSchedule(entry)}
-                        className="rounded p-1.5 text-muted-foreground hover:bg-muted hover:text-foreground focus:outline-none focus-visible:ring-2 focus-visible:ring-ring shrink-0"
-                        aria-label={`Edit ${entry.time}`}
-                      >
-                        <IconPencil size={14} stroke={1.5} />
-                      </button>
-                      {onDelete && entry.name !== undefined && (
-                        <DeleteButton
-                          name={entry.name}
-                          label={entry.time}
-                          onDelete={onDelete}
-                        />
-                      )}
-                    </li>
-                  );
-                })}
-              </ul>
-            )}
-          </section>
+          <ScheduleListView
+            entries={scheduleList}
+            togglingIds={togglingIds}
+            runningIds={runningIds}
+            onEdit={openEditSchedule}
+            onToggle={handleToggle}
+            onDelete={handleDelete}
+            onRunNow={
+              handleRunNow
+                ? (entry) => {
+                    handleRunNow(entry).catch(() => {});
+                  }
+                : undefined
+            }
+            onOpenDetails={onOpenDetails}
+          />
         )}
 
-        {scheduleViewMode === "calendar" &&
-          (() => {
-            const calendarSlots = buildCalendarTimeSlots(scheduleList);
-            return (
-              <section className="flex flex-col gap-8">
-                <div className="flex flex-col gap-2">
-                  <h3 className="text-xs font-semibold uppercase tracking-wide text-muted-foreground">
-                    Week view
-                  </h3>
-                  <div className="rounded-xl border border-border/70 bg-muted/20 overflow-hidden">
-                    <div className="grid grid-cols-8 text-sm">
-                      <div className="bg-muted/50 p-2 border-b border-r border-border/60 font-medium text-muted-foreground text-xs uppercase tracking-wider" />
-                      {WEEKDAY_LABELS.map((d, dayIndex) => (
-                        <div
-                          key={d}
-                          className={cn(
-                            "bg-muted/50 p-2 border-b border-border/60 font-medium text-muted-foreground text-center",
-                            dayIndex < WEEKDAY_LABELS.length - 1 &&
-                              "border-r border-border/60",
-                          )}
-                        >
-                          {d}
-                        </div>
-                      ))}
-                      {calendarSlots.map((timeLabel, timeIndex) => (
-                        <div key={timeLabel} className="contents">
-                          <div
-                            className={cn(
-                              "bg-muted/30 p-2 border-r border-border/60 text-muted-foreground text-xs flex items-center",
-                              timeIndex < calendarSlots.length - 1 &&
-                                "border-b border-border/60",
-                            )}
-                          >
-                            {timeLabel}
-                          </div>
-                          {WEEKDAY_LABELS.map((_, dayIndex) => {
-                            const entries = getEntriesInCell(
-                              scheduleList,
-                              dayIndex,
-                              timeLabel,
-                            );
-                            const isEmpty = entries.length === 0;
-                            const isLastRow =
-                              timeIndex === calendarSlots.length - 1;
-                            const isLastCol =
-                              dayIndex === WEEKDAY_LABELS.length - 1;
-                            return (
-                              <div
-                                key={`${timeLabel}-${dayIndex}`}
-                                className={cn(
-                                  "min-h-[52px] p-1.5 border-border/60 flex items-center justify-center",
-                                  !isLastCol && "border-r border-border/60",
-                                  !isLastRow && "border-b border-border/60",
-                                  isEmpty && "bg-background/50",
-                                )}
-                              >
-                                {isEmpty ? (
-                                  <span className="text-muted-foreground/40 text-xs">
-                                    —
-                                  </span>
-                                ) : (
-                                  <div className="w-full h-full min-h-[44px] rounded-lg p-1.5 flex flex-col gap-0.5 text-left">
-                                    {entries.map((entry) => (
-                                      <CalendarEntryPopover
-                                        key={entry.id}
-                                        entry={entry}
-                                        cellKey={`${dayIndex}-${timeLabel}`}
-                                        onEdit={openEditSchedule}
-                                      />
-                                    ))}
-                                  </div>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                </div>
-                {(() => {
-                  const loopEntries = scheduleList.filter(
-                    (e) =>
-                      e.enabled !== false &&
-                      e.time.match(/Every \d+ (minutes?|seconds?)/),
-                  );
-                  const onceEntries = scheduleList.filter(
-                    (e) => e.enabled !== false && e.time.startsWith("Once on"),
-                  );
-                  const monthlyEntries = scheduleList.filter(
-                    (e) =>
-                      e.enabled !== false && e.time.startsWith("Every month"),
-                  );
-                  if (
-                    loopEntries.length === 0 &&
-                    onceEntries.length === 0 &&
-                    monthlyEntries.length === 0
-                  ) {
-                    return null;
-                  }
-                  const sections: {
-                    title: string;
-                    entries: ScheduleEntry[];
-                  }[] = [
-                    { title: "Loop", entries: loopEntries },
-                    { title: "Monthly", entries: monthlyEntries },
-                    { title: "Once", entries: onceEntries },
-                  ];
-                  return (
-                    <div className="flex flex-col gap-8">
-                      {sections.map((section) =>
-                        section.entries.length > 0 ? (
-                          <div
-                            key={section.title}
-                            className="flex flex-col gap-1.5"
-                          >
-                            <h3 className="text-xs font-semibold uppercase tracking-wide text-muted-foreground">
-                              {section.title}
-                            </h3>
-                            <div className="flex flex-wrap gap-2">
-                              {section.entries.map((entry) => (
-                                <div
-                                  key={entry.id}
-                                  className="inline-flex items-center gap-2 rounded-lg border border-border/60 bg-muted/20 px-3 py-2 text-sm w-fit"
-                                >
-                                  <span className="text-foreground">
-                                    {entry.time}
-                                  </span>
-                                  <button
-                                    type="button"
-                                    onClick={() => openEditSchedule(entry)}
-                                    className="rounded p-1 text-muted-foreground hover:bg-muted hover:text-foreground"
-                                    aria-label={`Edit ${entry.time}`}
-                                  >
-                                    <IconPencil size={12} stroke={1.5} />
-                                  </button>
-                                </div>
-                              ))}
-                            </div>
-                          </div>
-                        ) : null,
-                      )}
-                    </div>
-                  );
-                })()}
-              </section>
-            );
-          })()}
+        {scheduleViewMode === "calendar" && (
+          <ScheduleCalendarView
+            entries={scheduleList}
+            onEdit={openEditSchedule}
+          />
+        )}
       </CardContent>
 
       <ScheduleFormDialog
@@ -965,6 +563,35 @@ export function ZeroScheduleCard({
         }
         saveError={saveError}
       />
+      <Dialog
+        open={pendingDelete !== null}
+        onOpenChange={(open) => {
+          if (!open) {
+            setPendingDelete(null);
+          }
+        }}
+      >
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Delete schedule?</DialogTitle>
+            <DialogDescription>
+              This will permanently delete the schedule{" "}
+              <span className="font-medium text-foreground">
+                {pendingDelete?.name}
+              </span>
+              . This action cannot be undone.
+            </DialogDescription>
+          </DialogHeader>
+          <DialogFooter>
+            <Button variant="outline" onClick={() => setPendingDelete(null)}>
+              Cancel
+            </Button>
+            <Button variant="destructive" onClick={confirmDelete}>
+              Delete
+            </Button>
+          </DialogFooter>
+        </DialogContent>
+      </Dialog>
     </Card>
   );
 }
