+++ turbo/apps/web/src/lib/schedule/index.ts
@@ -7,4 +7,5 @@ export {
   enableSchedule,
   disableSchedule,
   executeDueSchedules,
+  executeSchedule,
 } from "./schedule-service";
